requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1
django-appconf==1.0.6
django-ckeditor==6.7.1
django-compressor==4.5.1
django-crispy-forms==2.1
django-environ==0.11.2
django-htmx==1.17.2
django-js-asset==2.2.0
django-storages==1.14.4
django-tinymce==4.1.0
Faker==28.1.0
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
whitenoise==6.7.0
